Update capstone.py
Integrated prop and geometry writing so that everything can be written to a file. Fixed a bug with the theme only updating once. Added checkboxes to the Developer Insights menu for different types of geometry. Added more data validation for numeric input. Replaced ‘magic numbers’ for prop positioning and GUI spacing. Bugtested prefab writing (it has a lot of bugs), and redid some of the GUI alignment.